sun
The provided code snippet is a comment indicating a license agreement. The specific license mentioned is the "SUN LICENSE" and the link to the license is provided as "License http://www.sun.org/". This information is already present in the given code snippet. No additional code is required to answer this question.